requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
openai==1.13.3
pydantic==2.6.1
pydantic_core==2.16.2
pyTelegramBotAPI==4.15.4
python-dotenv==1.0.1
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.0
google-generativeai==0.4.0
mistralai==0.1.6
anthropic==0.19.2